Move requests implementations to a module
