+++ guillotina/request/__init__.py
@@ -1,0 +1,2 @@
+from .aiohttp import Request  # noqa
+from .asgi import GuillotinaRequest, GuillotinaWebSocket  # noqa
